+++ R/TwoSampleMR-package.R
@@ -1,14 +1,6 @@
-#' TwoSampleMR: Two Sample MR functions and interface to MR Base database
-#' 
-#' A package for performing Mendelian randomization using GWAS summary data. 
-#' It uses the [IEU GWAS database](https://gwas.mrcieu.ac.uk/) to obtain data automatically, 
-#' and a wide range of methods to run the analysis. You can use the [MR-Base web app](https://www.mrbase.org/) 
-#' to try out a limited range of the functionality in this package, 
-#' but for any serious work we strongly recommend using this R package.
-#' 
-#' **Full documentation available here:** [https://mrcieu.github.io/TwoSampleMR](https://mrcieu.github.io/TwoSampleMR/)
-#'
-#' @name TwoSampleMR-package
-#' @aliases TwoSampleMR twosamplemr
-#' @docType package
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
 NULL