+++ tests/testthat/test_add_metadata.r
@@ -1,5 +1,4 @@
 context("add metadata")
-library(TwoSampleMR)
 
 test_that("exposure data 1", {
 	d <- extract_instruments('ieu-a-2') %>% add_metadata()