+++ tests/testthat/test_eve.R
@@ -1,5 +1,4 @@
 context("eve")
-library(TwoSampleMR)
 
 dat <- make_dat("ieu-a-2", "ieu-a-7") %>% add_metadata()