+++ tests/testthat/test_outcomes.R
@@ -1,5 +1,4 @@
 context("outcome")
-library(TwoSampleMR)
 
 test_that("outcomes", {