+++ tests/testthat/test_singlesnp.R
@@ -1,5 +1,4 @@
 context("singlesnp")
-library(TwoSampleMR)
 
 dat <- make_dat("ieu-a-2", "ieu-a-7")